voice-pipeline/pyproject.toml
[project]
name = "voice-pipeline"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "sounddevice>=0.5",
  "numpy>=1.26",
  "onnxruntime>=1.17",
  "httpx>=0.27",
  "httpx-sse>=0.4",
  "openai>=1.30",
  "deepgram-sdk>=3.5",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "janus>=1.0",
]

[project.scripts]
voice-pipeline = "voice_pipeline.__main__:main"

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
]
